Files renamed to keep the same convention.
