+++ src/serializer.cpp
@@ -1,4 +1,4 @@
-#include "..//inc/Serializer.hpp"
+#include "..//inc/serializer.hpp"
 #include <vector>
 #include <fstream>
 #include <sstream>
