+++ tests/synchronizer_mock.hpp
@@ -1,4 +1,4 @@
-#include "..//inc/FileSynchronizer.hpp"
+#include "..//inc/file_synchronizer.hpp"
 #include "gmock/gmock.h" // Brings in gMock.
 
 class MockSynchronizer : public i_FileSynchronizer
